pyproject.toml
[project]
name = "mementoai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langfuse>=4.0.1",
    "litellm>=1.82.4",
    "llama-index-core>=0.14.18",
    "llama-index-vector-stores-mongodb>=0.10.1",
    "llama-index-storage-docstore-mongodb>=0.3.0",
    "llama-index-postprocessor-sbert-rerank>=0.3.0",
    "sentence-transformers>=3.0.0",
    "openai>=2.28.0",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "pymongo[srv]>=4.16.0",
    "python-multipart>=0.0.22",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
    "openinference-instrumentation-llama-index>=4.3.9",
    "llama-index-llms-litellm>=0.7.1",
    "llama-index-embeddings-litellm>=0.5.0",
    "langgraph>=1.1.3",
    "langchain-litellm>=0.6.2",
    "langchain>=1.2.13",
]
